moving z plane duplicate removal to decode step
